+++ src/main/java/org/sid/dao/TrainTravauxRepository.java
@@ -3,6 +3,6 @@ package org.sid.dao;
 import org.sid.entities.TrainsTravaux;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TrainTravauxRepository extends JpaRepository<TrainsTravaux, Long> {
+public interface TrainTravauxRepository extends JpaRepository<TrainsTravaux, Double> {
 
 }
